Docs: Add content to README
Chapters added with this commit: Developer´s goal, User´s goal
and an introduction to this application.
Saw also that 2 images in the testing chapter was not updated
with the new textcolor.
Links to spreadsheet and the application was redone
to look nicer.
Content menu was also added so it´s easier to navigate
through the README.